requirements.txt
torch>=2.0.0
einops
fire
Pillow
transformers
accelerate
sentencepiece
protobuf
huggingface_hub
safetensors
